requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
psutil>=5.9.0

# 可選依賴 (音樂功能)
# yt-dlp>=2023.0.0
# PyNaCl>=1.5.0
